Changed speed of game clock and adjust max character slots (#87)
* Changed speed of time passing to be accurate

Changed speed of in-game time from 30 sec of game time equaling 1 sec in the real world (30:1) to 48:1

* World Clock is still 10x as fast as it should be.

For some reason this formula is incrementing 10x faster than it should. Slow it down.

This should resolve #69

* Adjusted number of slots per account up to 8

The GUI in the .22 (Issue 0) version of the client can only clearly display 8 character slots. Adjusted this number up from 2 to 8, using the preexisting MAX_CHARACTER_SLOTS macro in Character.h. I couldn't figure out how to pull this from the config file, so I'll left a note for future commit.