Update databases
Add a `db_version` field to each database's `table_versions` to catalogue the version of the entire database.

Add a `salt` BLOB to the `segs` `accounts` table to store the store the salt string used in hashing the password.

Add an `options` and `gui` BLOB to the `segs_game` `accounts` table, and remove the `options` and `gui` tables themselves.