Entering the correct map based on character's map string
Given a correct map name in the character's data, and the proper data directory for the map.
The server will now create a MapInstance for the given map, and client will enter it.

TODO: needs a proper map scene graph to correctly set location of the entities after they enter the map.