Removed automatic admin and included text about adduser
Issue#408
I removed the automatic creation of admin when creating new databases or
forcing the creation of new databases. In it's place, use qInfo to
insert 3 lines of text. The first line warns the user that no admin
account was created. The next line lets the user know that to create an
admin, follow the schema on the next line. The final line is the schema
to follow to add a user. I have included two brief comments above this
small block listing what was changed.