For issue #302.
This is a lot of changes... To summarize, this is about removing m_mapName from CharData and creating m_current_map in EntityData. I might be making too much changes at one go here.

This is compilable but untested, I will be awaiting your inputs before going in further.
- Updated serializers for CharData and EntityData
- Incremented class versions for affected structs
- Character::from/toActualCharacter now takes in EntityData as parameter, where it is going to be serialized to/from QString
- Character::reset and getCurrentCostume no longer deals with map names
- Created uniquePtr for EntityData in Entity
- GameAccountResponseCharacterData now has m_serialized_entity_data, which will be used by GameHandler
- GameHandler::on_map_req now gets the map name from EntityData.m_current_map
- DataHelpers::getMapName and setMapName now takes in Entity as parameter

Some things to do off the top of my head:

- Test da ting (I feel a little scared to test this now tbh, with all the changes made :P)
- Fill up EntityData::reset and dump functions with something
- DataHelpers::getMapName is now unused, call it from somewhere else