Issue #474: Fixed crash after deleting a character and then creating a new one in the same slot.
The cause of this issue was that a copy of the character slot was reset, not the original one (in GameHandler::on_character_deleted).
The fix was to use a reference instead of a copy to access the data.

The same error was done in (GameHandler::on_delete_character), but it did not cause any damage as nothing were written to the slot here.
I changed the copy to a (const) reference here as well to reduce the risk of someone doing the same mistake again.