Creating a Semi-Standalone EmailHandler that derives from EventProcessor (Part 2) (#548)
Great PR @eos428 !!! Thank you for tackling this issue

* Closes #367.

A very barebones implementation of EmailHandler at the moment. Further details are in PR opening post.

* Deleted Common/Email folders and its contents placed into Servers/MapServer/Events folder instead.
EmailHandler is usable, but only for email headers.
EmailHandler is initialized in MapServer (not sure if the right place... feel free to let me know)
Changed up the email functions in DataHelpers for testing purposes. Should we remove these DataHelpers functions and use the EmailHandler directly from the Cmd commands?

* Some update to the email handler. Now incorporates a message bus endpoint that subscribes to InternalEvents::onClientConnected and onClientDisconnected.
TODO: Investigate why the handler becomes unusable when two events are fired to it at once (when opening the email tab for the first time, the client will send both /emailheaders and /emailread 152)

* Should be close to a working EmailHandler now (without db involved yet).
When gameHandler is given evClientConnected and evClientDisconnected in its queue, it will post those events globally to the message bus for others (such as EmailHandler) to listen

* EmailHandler and EmailEvents moved to GameServer
There's a lot of commented code in EmailHandler.cpp (the ones where EmailHeaders and EmailRead are declared especially) but they shouldn't be there in the first place, and should be done in MapInstance instead. I will continue on by making the EmailHandler return a response to the MapInstance, and then the header/email read is declared there instead.

* Moved EmailHandler and EmailEvents to GameServer
A lot of stuff I will explain in the PR comment instead

* Giving EmailHandler access to GameDbSyncHandler (we'll need it because this is related to DB stuff)

* Update Email Events

* build ruuuuuuuuuuuuns

* Changes to things carried inside EmailEvents and ready to test without DB for now.

Current problem is that I putq'd EmailHeaderRequest() but it doesn't to be received properly by the email handler, and it even clogs the EmailHandler with it not able to receive any more events.

* Sending email headers to self works - Good step forward!

* EmailRead and EmailWasReadByRecipient events seem to work now

- Added mock email database as a map
- Separated EmailRead into Request and Response

* Preparation for email serialization and putting them in db

* More preparation for email integration with database.

Have some questions that I'll post in github PR comments instead of here.

* A lot of updates, but we are not there yet.

Changes made:
- Start to fill up the email funcs in GameDbSyncContext
- m_email_id will be given from the db (auto-incrementation there)
- GameDbSyncHandler funcs should be okay
- Email timestamp changed to uint32_t, I think this should be created in db as well

* Connected email funcs in GameDbSyncContext and GameDbSyncHandler in places where they are needed

- Split up email_update to email_mark_as_read and email_update_on_char_delete
- Maybe don't need to use getEmailBySender/RecipientId
- Still have to parse /emailsend!

* A lot of progress done

- EmailRead and the SQL queries related to it are working properly
- EmailHeaders will be unused since the test func is to send email to self
- m_prepared_email_insert is working properly
- Can parse /emailsend params, but untested if quotation marks are used in email recipient name, subject, and/or content
- Added EmailDefinitions.h
- Increased required_db_version in GameDbSyncContext
- All new SQL queries are not throwing any errors now (thanks dracc for the help)

* ready for testing

* Fixes for broxen's issues on testing

* ClientDisconnectedMessage should send correct char db id now