Most of Nemerle's changes
All of Nemerle's changes are done execpt for two functions.

giveEnhancement
contact_dialog

I have Lua sending in tables as it simplifies things on the scripting side.
So there is some conversion that needs to happen.

Also, there is an issue (Broxen is aware of) with referencing functions within the ScriptingEngine in the registered types.
Thus most of the larger functions have moved to DataHelpers.
I could move two that have Lua tables being sent but that would result in a reference to Sol2.hpp in datahelpers.