Updating version for -dev build
Releases should have their own fixed version number, so we can tell from the logs if someone is using a release package or a dev build. Would be great if there were a way to append build number so we know what build someone is running from github. I can't think of a way to automate that easily.