Updating version for -dev build (#798)
* Updating version for -dev build

Releases should have their own fixed version number, so we can tell from the logs if someone is using a release package or a dev build. Would be great if there were a way to append build number so we know what build someone is running from github. I can't think of a way to automate that easily.

* Git rev additions

Added variables
- SEGS_VERSION_MAJOR
- SEGS_VERSION_MINOR
- SEGS_VERSION_PATCH
- SEGS_DESCRIBE
- SEGS_REVISION
- SEGS_BRANCH

* Use cmake variable for version

* Fix typo in SEGS_DESCRIPTION

* Removed unnecessary pragmas and fixed Version from cmake

* [skip ci] formatting changes