Fixes, clean-up and adjustments
Spawndefs moved to map-specific subdirectories to resolve the file system permission issue on Linux and reorganize for future additions to the system. Additionally, a fix to a bizarre and previously unknown issue with the iteration of map directories was resolved. (Thanks, Nemerle!)

Adjusted some spawndefs to improve variety and performance to encounter generation for some maps.

Co-Authored-By: ZopharShinta <zopharshinta@users.noreply.github.com>
Co-Authored-By: Artur K. <nemerle5 dwelling_at gmail.com>
Co-Authored-By: ItinerantEngineer <itinerantengineer@users.noreply.github.com>