Renamed SE: Null to SE: Template; Adjusted base timers
ScriptingEngine_Null has been renmaed to ScriptingEngine_Template as a way to reserve it for possible future use. In case a new or additional scripting engine is employed. it may act as a good starting skeleton.

Base refresh times for encounters increased to 2 to 3 minutes for most maps. This leaves more time for interaction while allowing maps to stay fresh for exploration.