Categorize the "Sending new Powerstance" message
[skip ci] 
This message fires too frequently, because toggle powers are constantly sending stance. It clutters the logs and makes them hard to follow on Red/Blue
We may want to actually change that behavior to only send on activation/deactivation, but for now, let's move this to Animations where it makes more sense and can be toggled.